fixed golden bat poop to reach integer limit, just uses long instead

+++ Utils/RagnarokModPlayer.cs
@@ -251,10 +251,10 @@ namespace RagnarokMod.Utils
 					// is it a coin?
 					if(base.Player.inventory[i].type >= 71 && base.Player.inventory[i].type <= 74) 
 					{
-						effectivecoinvalue += (int)(Math.Pow( 100, base.Player.inventory[i].type - 71) * base.Player.inventory[i].stack);
+						effectivecoinvalue += (long)(Math.Pow( 100, base.Player.inventory[i].type - 71) * base.Player.inventory[i].stack);
 					}
 				}
-				float damagemodifier = (float)Math.Pow(((float)effectivecoinvalue / 10000000), 0.2) / 10;
+				float damagemodifier = (float)Math.Pow(((float)(effectivecoinvalue / 10000000)), 0.2) / 10;
 				base.Player.GetDamage(DamageClass.Generic) += damagemodifier;	
 			}
